Big refactor, add adaptives
