+++ src/pages/DashboardPage/components/Card/Card.module.less
@@ -35,6 +35,10 @@
   align-items: center;
   justify-content: space-between;
   padding: 8px;
+
+  & > div:last-child {
+    align-items: flex-end;
+  }
 }
 .borrowCardStatContainer {
   gap: unset;
@@ -49,6 +53,10 @@
   display: flex;
   flex-direction: column;
   padding: 8px;
+
+  & > div:last-child {
+    line-height: 21px;
+  }
 }
 .lendCardBadge {
   display: flex;
