+++ src/pages/DashboardPage/components/DashboardBorrowTab/DashboardBorrowTab.module.less
@@ -1,19 +1,3 @@
-.container {
-  display: flex;
-  padding: 0 24px;
-  gap: 24px;
-  width: 100%;
-
-  @media (max-width: 1240px) {
-    flex-direction: column;
-  }
-
-  @media (max-width: 640px) {
-    padding: 0 16px;
-    gap: 16px;
-  }
-}
-
 .nftsSection,
 .additionalContentSection {
   width: 50%;
@@ -33,11 +17,19 @@
 
   @media (max-width: 1240px) {
     display: flex;
+
+    & > div {
+      min-width: 180px;
+    }
   }
 
   @media (max-width: 640px) {
     padding-top: 8px;
     gap: 16px;
+
+    & > div {
+      min-width: 154px;
+    }
   }
 }
