+++ src/pages/DashboardPage/components/DashboardLendTab/DashboardLendTab.module.less
@@ -1,19 +1,3 @@
-.connectedContainer {
-  display: flex;
-  padding: 0 24px;
-  gap: 24px;
-  width: 100%;
-
-  @media (max-width: 1240px) {
-    flex-direction: column;
-  }
-
-  @media (max-width: 640px) {
-    padding: 0 16px;
-    gap: 16px;
-  }
-}
-
 .collectionsSection,
 .additionalContentSection {
   width: 50%;
@@ -23,22 +7,102 @@
   }
 }
 
+.collectionsSection.fullWidth {
+  width: 100%;
+}
+
 .collectionsCardList {
   display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(162px, 2fr));
+  grid-template-columns: repeat(auto-fill, minmax(180px, 2fr));
   padding-top: 12px;
   gap: 24px;
 
   @media (max-width: 1240px) {
     display: flex;
+
+    & > div {
+      min-width: 180px;
+    }
   }
 
   @media (max-width: 640px) {
     padding-top: 8px;
     gap: 16px;
+
+    & > div {
+      min-width: 154px;
+    }
   }
 }
 
 .notConnectedContainer {
-  padding: 0 24px;
+  width: 100%;
+}
+
+.allTimeContent {
+  display: flex;
+  align-items: center;
+  margin-bottom: 24px;
+  gap: 16px;
+  width: 100%;
+}
+.allTimeStatsContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 16px;
+  min-height: 188px;
+  height: 100%;
+  width: 100%;
+}
+.allTimeStats {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+}
+.allTimeChartStats {
+  display: flex;
+  flex-direction: column;
+
+  & > div:nth-child(odd) {
+    background: var(--bg-primary-deep);
+  }
+}
+
+.allocationContent {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 24px;
+}
+.allocationStatsContainer {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  min-height: 160px;
+  height: 100%;
+  width: 100%;
+}
+.allocationStats {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+}
+.allocationChartStats {
+  display: flex;
+  flex-direction: column;
+
+  & > div:nth-child(odd) {
+    background: var(--bg-primary-deep);
+  }
+}
+
+.doughnutChart {
+  width: 144px;
+  height: 144px;
+}
+
+.manageOffersButton {
+  margin-top: 16px;
+  width: 100%;
 }
