--- src/pages/DashboardPage/components/DashboardLendTab/components/AllTimeBlock/AllTimeBlock.module.less
@@ -1,45 +1,0 @@
-.allTimeHeading {
-  display: flex;
-  align-items: center;
-  padding: 24px 0 12px;
-  font: var(--title-sm);
-  text-transform: uppercase;
-  line-height: 32px;
-}
-
-.allTimeContent {
-  display: flex;
-  align-items: center;
-  margin-bottom: 24px;
-  gap: 16px;
-  width: 100%;
-}
-
-.allTimeStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 16px;
-  min-height: 188px;
-  height: 100%;
-  width: 100%;
-}
-
-.allTimeStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.allTimeChartStats {
-  display: flex;
-  flex-direction: column;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-primary-deep);
-  }
-}
-
-.manageOffersButton {
-  width: 100%;
-}
