--- src/pages/DashboardPage/components/DashboardLendTab/components/AllocationBlock/AllocationBlock.module.less
@@ -1,49 +1,0 @@
-.allocationHeading {
-  display: flex;
-  align-items: center;
-  padding: 24px 0 12px;
-  font: var(--title-sm);
-  text-transform: uppercase;
-  line-height: 32px;
-}
-
-.allocationContent {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: 24px;
-}
-
-.allocationStatsContainer {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  min-height: 160px;
-  height: 100%;
-  width: 100%;
-}
-
-.allocationStats {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.allocationChartStats {
-  display: flex;
-  flex-direction: column;
-
-  & > div:nth-child(odd) {
-    background: var(--bg-primary-deep);
-  }
-}
-
-.manageOffersButton {
-  margin-top: 16px;
-  width: 100%;
-}
-
-.doughnutChart {
-  width: 144px;
-  height: 144px;
-}
