--- src/pages/DashboardPage/components/DashboardLendTab/components/AllocationBlock/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AllocationBlock'
