+++ src/pages/DashboardPage/components/components.tsx
@@ -13,14 +13,12 @@ interface SearchableHeadingProps<T> {
 export const SearchableHeading = <T extends SearchableHeadingProps<T>>({
   title,
   searchSelectParams,
-}: SearchableHeadingProps<T>) => {
-  return (
-    <div className={styles.searchableHeadingWrapper}>
-      <h4 className={styles.searchableHeading}>{title}</h4>
-      <SearchSelect className={styles.searchSelect} {...searchSelectParams} />
-    </div>
-  )
-}
+}: SearchableHeadingProps<T>) => (
+  <div className={styles.searchableHeadingWrapper}>
+    <h4 className={styles.searchableHeading}>{title}</h4>
+    <SearchSelect className={styles.searchSelect} {...searchSelectParams} />
+  </div>
+)
 
 export const Heading = ({ title }: { title: string }) => <h4 className={styles.heading}>{title}</h4>
 
@@ -30,17 +28,15 @@ interface ChartStatInfoProps {
   indicatorColor: string
 }
 
-export const ChartStatInfo: FC<ChartStatInfoProps> = ({ value, label, indicatorColor }) => {
-  return (
-    <div className={styles.chartStatInfo}>
-      <span className={styles.chartStatInfoValue}>{value}</span>
-      <div className={styles.chartStatInfoLabelWrapper}>
-        <span className={styles.chartStatInfoLabel}>{label}</span>
-        <span style={{ background: indicatorColor }} className={styles.chartStatInfoIndicator} />
-      </div>
+export const ChartStatInfo: FC<ChartStatInfoProps> = ({ value, label, indicatorColor }) => (
+  <div className={styles.chartStatInfo}>
+    <span className={styles.chartStatInfoValue}>{value}</span>
+    <div className={styles.chartStatInfoLabelWrapper}>
+      <span className={styles.chartStatInfoLabel}>{label}</span>
+      <span style={{ background: indicatorColor }} className={styles.chartStatInfoIndicator} />
     </div>
-  )
-}
+  </div>
+)
 
 export const DashboardStatInfo: FC<StatsInfoProps> = (props) => {
   const { classNamesProps, ...rest } = props
